Spacemouse platform dependent code is moved to different classes. Compiles and works for Linux
