+++ src/Gui/GuiApplicationNativeEventAware.h
@@ -29,30 +29,12 @@
 
 class QMainWindow;
 
-#if defined(Q_OS_WIN) && defined(_USE_3DCONNEXION_SDK)
-#include "3Dconnexion/MouseParameters.h"
-
-#include <vector>
-#include <map>
-
-//#define _WIN32_WINNT 0x0501  //target at least windows XP
-#include <Windows.h>
-#include <QAbstractNativeEventFilter>
-
-#elif defined(Q_OS_MACX) && defined(_USE_3DCONNEXION_SDK)
-
-#include <IOKit/IOKitLib.h>
-#include <ConnexionClientAPI.h>
-// Note that InstallConnexionHandlers will be replaced with
-// SetConnexionHandlers "in the future".
-extern OSErr InstallConnexionHandlers(ConnexionMessageHandlerProc messageHandler,
-                                      ConnexionAddedHandlerProc addedHandler,
-                                      ConnexionRemovedHandlerProc removedHandler)
-                                      __attribute__((weak_import));
-extern UInt16 RegisterConnexionClient(UInt32 signature, UInt8 *name, UInt16 mode,
-                                      UInt32 mask) __attribute__((weak_import));
-extern void UnregisterConnexionClient(UInt16 clientID) __attribute__((weak_import));
-extern void CleanupConnexionHandlers(void) __attribute__((weak_import));
+#if defined(Q_OS_LINUX)
+#include "3Dconnexion/GuiNativeEventLinux.h"
+#elif defined(Q_OS_WIN)
+#include "3Dconnexion/GuiNativeEventWin32.h"
+#elif defined(Q_OS_MACX)
+#include "3Dconnexion/GuiNativeEventMac.h"
 #endif // Platform switch
 
 namespace Gui
@@ -65,98 +47,15 @@ namespace Gui
         ~GUIApplicationNativeEventAware();
         void initSpaceball(QMainWindow *window);
         bool isSpaceballPresent() const {return spaceballPresent;}
+        void setSpaceballPresent(bool present) {spaceballPresent = present;} 
         bool processSpaceballEvent(QObject *object, QEvent *event);
     private:
         bool spaceballPresent;
-        QMainWindow *mainWindow;
         int  motionDataArray[6];
         bool setOSIndependentMotionData();
         void importSettings();
         float convertPrefToSensitivity(int value);
-
-#ifdef Q_OS_LINUX
-    public:
-        void pollSpacenav();
-#endif // Q_OS_LINUX
-
-#ifdef _USE_3DCONNEXION_SDK
-// For Windows
-#ifdef Q_OS_WIN
-    class RawInputEventFilter : public QAbstractNativeEventFilter
-    {
-    public:
-        typedef bool (*EventFilter)(void *message, long *result);
-        RawInputEventFilter(EventFilter filter) : eventFilter(filter) {
-        }
-        virtual ~RawInputEventFilter() {
-        }
-
-        virtual bool nativeEventFilter(const QByteArray & /*eventType*/, void *message, long *result) {
-            return eventFilter(message, result);
-        }
-
-    private:
-        EventFilter eventFilter;
-    };
-
-    public:
-        static bool Is3dmouseAttached();
-
-        I3dMouseParam& MouseParams();
-        const I3dMouseParam& MouseParams() const;
-
-        virtual void Move3d(HANDLE device, std::vector<float>& motionData);
-        virtual void On3dmouseKeyDown(HANDLE device, int virtualKeyCode);
-        virtual void On3dmouseKeyUp(HANDLE device, int virtualKeyCode);
-
-    private:
-        bool InitializeRawInput(HWND hwndTarget);
-        static bool RawInputEventFilter(void* msg, long* result);
-        void OnRawInput(UINT nInputCode, HRAWINPUT hRawInput);
-        UINT GetRawInputBuffer(PRAWINPUT pData, PUINT pcbSize, UINT cbSizeHeader);
-        bool TranslateRawInputData(UINT nInputCode, PRAWINPUT pRawInput);
-        bool ParseRawInput(UINT nInputCode, PRAWINPUT pRawInput);
-        void On3dmouseInput();
-
-        class TInputData
-        {
-        public:
-            TInputData() : fAxes(6) {}
-
-            bool IsZero() {
-                return (0.==fAxes[0] && 0.==fAxes[1] && 0.==fAxes[2] &&
-                        0.==fAxes[3] && 0.==fAxes[4] && 0.==fAxes[5]);
-            }
-
-            int fTimeToLive; // For telling if the device was unplugged while sending data
-            bool fIsDirty;
-            std::vector<float>     fAxes;
-        };
-
-        HWND fWindow;
-
-        // Data cache to handle multiple rawinput devices
-        std::map< HANDLE, TInputData>       fDevice2Data;
-        std::map< HANDLE, unsigned long>    fDevice2Keystate;
-        // 3dmouse parameters
-        MouseParameters f3dMouseParams;     // Rotate, Pan Zoom etc.
-        // use to calculate distance traveled since last event
-        DWORD fLast3dmouseInputTime;
-        static Gui::GUIApplicationNativeEventAware* gMouseInput;
-#endif // Q_OS_WIN
-#ifdef Q_OS_MACX
-    private:
-        static UInt16 tdxClientID; /* ID assigned by the driver */
-        static uint32_t lastButtons;
-    public:
-        static void tdx_drv_handler( io_connect_t connection,
-                                     natural_t messageType,
-                                     void *messageArgument );
-        void Move3d();
-        void Button3d(bool buttonDown, int buttonNumber);
-
-#endif// Q_OS_MACX
-#endif // _USE_3DCONNEXION_SDK
+        GuiNativeEvent *nativeEvent;
     }; // end class GUIApplicationNativeEventAware
 } // end namespace Gui
